RDMA/hns: Fix possible RAS when DCA is not attached
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB30V8

----------------------------------------------------------------------

RAS may occur if the DCA buffer is not attached and the DB
is knocked out.

This patch adds a safe page for DCA, which will be attached
to QP if no DCA buffer is attached to avoid the HW accessing
illegal addresses.

Fixes: 10bb3b802412 ("RDMA/hns: Add method for attaching WQE buffer")
Signed-off-by: Chengchang Tang <tangchengchang@huawei.com>
Signed-off-by: Xinghai Cen <cenxinghai@h-partners.com>