migration: modify dfx error type without VM driver
driver inclusion
category: cleanup
bugzilla: https://gitee.com/openeuler/kernel/issues/IB9H0P
CVE: NA

----------------------------------------------------------------------

In the live migration scenario, when the ACC driver is not insmod
to the VM, the command to test the mailbox operation will detect
that the device has no driver loaded. However, the error reported
is the same as the abnormal parameter when testing dfx, and the
error scenario cannot be distinguished.
Therefore, this error type needs to be modified.

Signed-off-by: Longfang Liu <liulongfang@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>