crypto: hisilicon/sec2 - fix for aead icv error
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IBATCK
CVE: NA

----------------------------------------------------------------------

When the AEAD algorithm is used for encryption or decryption,
the input authentication length varies, the hardware needs to
obtain the input length to pass the integrity check verification.
Currently, the driver uses a fixed authentication length,which
causes decryption failure, so the length configuration is modified.
In addition, the step of setting the auth length is unnecessary,
so it was deleted from the setkey function.

Fixes: 2f072d75d1ab ("crypto: hisilicon - Add aead support on SEC2")
Signed-off-by: Wenkai Lin <linwenkai6@hisilicon.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>