fscache: modify the waiting mechanism with duplicate volumes
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB5UKT

--------------------------------

In the current erofs ondemand scenario, fd holds a reference to object,
object holds a reference to cookie, and cookie holds a reference to
volume. The cookie->ref is subtracted when the object->ref reaches 0. And
only when cookie->ref reaches 0 will volume->ref be decremented by one. It
means a dependency chain is formed between fd->object->cookie->volume.

The unhashing of the cookie does not depend on its reference count being
zero; it only requires its state to be FSCACHE_COOKIE_STATE_RELINQUISHING.

But for the volume, the situation is different. Volume will be unhashed
only when its reference count is zero. That means it depends on the user
space closing the fd of the cachefiles_object. If the relinquish process
is finished but the user-space fails to properly close the fd, it will
cause the kernel to hang in fscache_wait_on_volume_collision(), which is
an inappropriate behavior.

Modify the waiting mechanism to be interruptible to fix this problem.

Fixes: 62ab63352350 ("fscache: Implement volume registration")
Signed-off-by: Zizhi Wo <wozizhi@huawei.com>