PM / devfreq: event: Supplement the name content of the event in devfreq-event
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBC4SJ

--------------------------------

The name of the event in devfreq-event only has the ID as a distinguishing
mark, which is too simple and not intuitive enough to distinguish different
events.

Therefore, the name parameter of desc is added to the name of the event.

Signed-off-by: Xiangwei Li <liwei728@huawei.com>