RDMA/hns: Fix creating GSI QP in non-extended SGE QP bank
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB71IG

----------------------------------------------------------------------

According to the IB protocol, the qp_num of the GSI QP must
be set to 1. Therefore, the QP must be created in BANK 1.

Currently, only QPs in BANK 0 and 6 can use extended SGEs,
but the GSI QP in bank 1 also need to use extended SGEs.
To fix it, the restriction is changed to allow QPs in BANK
1 and BANK 6 to use extended SGEs.

Fixes: 209f40403456 ("RDMA/hns: Fix RoCEE hang when multiple QP banks use EXT_SGE EXT_SGE")
Signed-off-by: wenglianfa <wenglianfa@huawei.com>
Signed-off-by: Xinghai Cen <cenxinghai@h-partners.com>