dts: bcm2711: Don't mark timer regs unconfigured
The DT property arm,cpu-registers-not-fw-configured tells the kernel
that the ARM architectural timer has not been configured by the
firmware. This prevents the use of a vDSO - a faster alternative to a
syscall for some common kernel operations.

However, on Pi 4 the firmware does configure the timer, so this property
is unnecessary. Delete it.

Also remove it from the bcm2712.dtsi where it should never have been,
since it is only relevant to 32-bit ARM kernels.

Signed-off-by: Phil Elwell <phil@raspberrypi.com>